# frontend/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ShowtimeBackend | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5146/"
    }
  ],
  "paths": {
...
    "/api/admin/seat-maps/{seatMapId}/sections": {
      "get": {
        "tags": [
          "Seat Zone Administration - Seat Sections"
        ],
        "parameters": [
          {
            "name": "seatMapId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "SectionType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsSellable",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPagedResponseOfSeatSectionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPagedResponseOfSeatSectionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPagedResponseOfSeatSectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Seat Zone Administration - Seat Sections"
        ],
        "parameters": [
          {
            "name": "seatMapId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/admin/seat-sections/{seatSectionId}": {
      "get": {
        "tags": [
          "Seat Zone Administration - Seat Sections"
        ],
        "parameters": [
          {
            "name": "seatSectionId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Seat Zone Administration - Seat Sections"
        ],
        "parameters": [
          {
            "name": "seatSectionId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeatSectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfSeatSectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Seat Zone Administration - Seat Sections"
        ],
        "parameters": [
          {
            "name": "seatSectionId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfPagedResponseOfSeatSectionResponse": {
        "required": [
          "success",
          "data",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PagedResponseOfSeatSectionResponse"
              }
            ]
          },
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseOfSeatSectionResponse": {
        "required": [
          "success",
          "data",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SeatSectionResponse"
              }
            ]
          },
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PagedResponseOfSeatSectionResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatSectionResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "SeatSectionRequest": {
        "required": [
          "sectionCode",
          "sectionName",
          "sectionType",
          "sectionColor",
          "floorNo",
          "isSellable",
          "displayOrder",
          "remark"
        ],
        "type": "object",
        "properties": {
          "sectionCode": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "sectionType": {
            "type": "string"
          },
          "sectionColor": {
            "type": [
              "null",
              "string"
            ]
          },
          "floorNo": {
            "type": [
              "null",
              "string"
            ]
          },
          "isSellable": {
            "type": "boolean"
          },
          "displayOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "remark": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SeatSectionResponse": {
        "required": [
          "seatSectionId",
          "seatMapId",
          "sectionCode",
          "sectionName",
          "sectionType",
          "sectionColor",
          "floorNo",
          "isSellable",
          "displayOrder",
          "remark"
        ],
        "type": "object",
        "properties": {
          "seatSectionId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "seatMapId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "sectionCode": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "sectionType": {
            "type": "string"
          },
          "sectionColor": {
            "type": [
              "null",
              "string"
            ]
          },
          "floorNo": {
            "type": [
              "null",
              "string"
            ]
          },
          "isSellable": {
            "type": "boolean"
          },
          "displayOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "remark": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Provide the JWT access token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}